+++ worker/tasks/dataflux_func/runner.py
@@ -25,7 +25,7 @@ from worker.tasks import BaseResultSavingTask
 
 # Current Module
 from worker.tasks import BaseTask
-from worker.tasks.dataflux_func import DataFluxFuncBaseException, ScriptNotFoundException, FunctionNotFoundException
+from worker.tasks.dataflux_func import DataFluxFuncBaseException, NotFoundException, NotFoundException
 from worker.tasks.dataflux_func import ScriptBaseTask
 
 CONFIG = yaml_resources.get('CONFIG')
@@ -292,7 +292,7 @@ def dataflux_func_runner(self, *args, **kwargs):
         target_script = SCRIPT_DICT_CACHE.get(script_id)
 
         if not target_script:
-            e = ScriptNotFoundException('Script `{}` not found'.format(script_id))
+            e = NotFoundException('Script `{}` not found'.format(script_id))
             raise e
 
         extra_vars = {
@@ -321,7 +321,7 @@ def dataflux_func_runner(self, *args, **kwargs):
         # 执行脚本
         entry_func = script_scope.get(func_name)
         if not entry_func:
-            e = FunctionNotFoundException('Function `{}` not found in `{}`'.format(func_name, script_id))
+            e = NotFoundException('Function `{}` not found in `{}`'.format(func_name, script_id))
             raise e
 
         # 执行函数